+++ features/step_definitions/rabbitmq_specific_steps.rb
@@ -1,0 +1,5 @@
+Then 'I expect all the publishes to have been acknowledged' do
+  ctx = test_runner.fetch_current_adapter_context
+  expect(ctx.channel).to be_using_publisher_confirms
+  expect(ctx.channel.unconfirmed_set).to be_empty
+end